// packages/kratos/src/lib/kratos/models/UiNode.ts
export interface UiNode {
    /**
     * 
     * @type {UiNodeAttributes}
     * @memberof UiNode
     */
    attributes: UiNodeAttributes;
    /**
     * Group specifies which group (e.g. password authenticator) this node belongs to.
     * default DefaultGroup
     * password PasswordGroup
     * oidc OpenIDConnectGroup
     * profile ProfileGroup
     * link LinkGroup
     * code CodeGroup
     * totp TOTPGroup
     * lookup_secret LookupGroup
     * webauthn WebAuthnGroup
     * passkey PasskeyGroup
     * identifier_first IdentifierFirstGroup
     * @type {string}
     * @memberof UiNode
     */
    group: UiNodeGroupEnum;
    /**
     * 
     * @type {Array<UiText>}
     * @memberof UiNode
     */
    messages: Array<UiText>;
    /**
     * 
     * @type {UiNodeMeta}
     * @memberof UiNode
     */
    meta: UiNodeMeta;
    /**
     * The node's type
     * text Text
     * input Input
     * img Image
     * a Anchor
     * script Script
     * @type {string}
     * @memberof UiNode
     */
    type: UiNodeTypeEnum;
}


/**
 * @export
 */
export const UiNodeGroupEnum = {
    Default: 'default',
    Password: 'password',
    Oidc: 'oidc',
    Profile: 'profile',
    Link: 'link',
    Code: 'code',
    Totp: 'totp',
    LookupSecret: 'lookup_secret',
    Webauthn: 'webauthn',
    Passkey: 'passkey',
    IdentifierFirst: 'identifier_first'
} as const;
export type UiNodeGroupEnum = typeof UiNodeGroupEnum[keyof typeof UiNodeGroupEnum];

/**
 * @export
 */
export const UiNodeTypeEnum = {
    Text: 'text',
    Input: 'input',
    Img: 'img',
    A: 'a',
    Script: 'script'
} as const;
export type UiNodeTypeEnum = typeof UiNodeTypeEnum[keyof typeof UiNodeTypeEnum];